move student components to student folder
