+++ s-with-me-front/src/components/student/libarary/LibraryApp.jsx
@@ -3,7 +3,7 @@ import AppLayout from '../AppLayout';
 import BookOverview from './BookOverview';
 import LibraryFolderList from './LibraryFolderList';
 
-import { withStyles, css, withStylesPropTypes } from '../../common-ui/withStyles';
+import { withStyles, css, withStylesPropTypes } from '../../../common-ui/withStyles';
 
 class LibraryApp extends PureComponent {
   static propTypes = {
