+++ s-with-me-front/src/components/student/libarary/LibraryFolderList.jsx
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 
-import Card from '../../common-ui/Card';
+import Card from '../../../common-ui/Card';
 import LibraryFolderFilter from './LibraryFolderFilter';
 import LibraryFolderTable from './LibraryFolderTable';
