+++ s-with-me-front/src/components/student/libarary/LibraryFolderTable.jsx
@@ -1,10 +1,10 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
-import Table from '../../common-ui/table/Table';
-import TableBody from '../../common-ui/table/TableBody';
-import TableRow from '../../common-ui/table/TableRow';
-import TableCell from '../../common-ui/table/TableCell';
+import Table from '../../../common-ui/table/Table';
+import TableBody from '../../../common-ui/table/TableBody';
+import TableRow from '../../../common-ui/table/TableRow';
+import TableCell from '../../../common-ui/table/TableCell';
 
 export default class LibraryFolderTable extends PureComponent {
   static propTypes = {
