+++ s-with-me-front/src/stories/BookOverviewStory.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import BookOverview from '../components/libarary/BookOverview';
+import BookOverview from '../components/student/libarary/BookOverview';
 
 storiesOf('Library/BookOverview', module).addWithJSX('overview default', () => <BookOverview />);
