+++ s-with-me-front/src/stories/LibraryAppStory.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import LibraryApp from '../components/libarary/LibraryApp';
+import LibraryApp from '../components/student/libarary/LibraryApp';
 
 storiesOf('Library/LibraryApp', module).addWithJSX('Library App', () => <LibraryApp />);
